==> 01-Kayak/01_cities_weather.ipynb <==
# Weather information: Top 35 cities to visit in France  

In this notebook I :  
-  Retrieved the weather information for the [top 35 best cities in France](https://one-week-in.com/35-cities-to-visit-in-france/)
-  Defined the top 5 cities to visit based on the best weather for the next 7 days

## Import libraries

import requests
import pandas as pd
import numpy as np
import json

from src.cities_weather import req_gps_fn, req_weather_fn, req_to_df_fn
from src.cities_weather import explode_df_fn, explode2_df_fn, explodeall_df_fn

## Import data

# Import list of top 35 cities in France
filename = 'data/raw/top_35_cities_france.txt'
with open(filename) as file:
    top_cities_ls = [line.rstrip() for line in file]

print(len(top_cities_ls))
print(top_cities_ls[0:4])

## 1. GPS coordinates

Use the [nominatim API](https://nominatim.org/) to get latitute and longitude for each of the cities of interest 


### API request

# Use previously defined requests function  to retrieve response for top cities
api_gps=[req_gps_fn(x) for x in top_cities_ls]
print('API response length: ', len(api_gps))
api_gps[0]

# Create dataframe from request response
api_cities_df = req_to_df_fn(api_gps)
api_cities_df.head()

# Keep copy of original df
api_cities_df_raw=api_cities_df
api_cities_df=api_cities_df_raw.copy()

### Delete duplicated alias

#Create a column containing search name of city (obtained from display_name column)
api_cities_df['Search_Name']=api_cities_df.display_name.str.split(',',expand=True)[0]
api_cities_df.head()

#Sort by entry with highest importance (given based on Wikipedia entries) and then delete duplicated Search Names
api_cities_df=api_cities_df.sort_values('importance', ascending=False).drop_duplicates(subset=['Search_Name'])
api_cities_df=api_cities_df.sort_values('importance', ascending=False).drop_duplicates(subset=['Search_Name'])
print('Shape of duplicate clean cities dataframe: ', api_cities_df.shape)
api_cities_df.head()

### GPS coordinates dataframe

#The GPS coordinates are found under the keys lat and lon
cols_keep_ls=["Search_Name","lat","lon"]
GPS_df=api_cities_df.loc[:,cols_keep_ls]

print('Cities GPS dataframe shape: ', GPS_df.shape)
GPS_df.head(44)

#When filtering df by  names from initial list of cities we loose some of them. Update name in top cities file and re-do API request
s=pd.Series(top_cities_ls)
print(s[~s.isin(GPS_df["Search_Name"])])

# Final GPS info dataframe. Select search names in list of top cities
GPS_df=GPS_df[GPS_df["Search_Name"].isin(top_cities_ls)]
GPS_df.sort_values(by=['Search_Name'],inplace=True)
print(GPS_df.shape)
GPS_df.head()

## 2. Weather data  
Use the [openweathermap API](https://openweathermap.org/api/one-call-api) and previously obtained GPS coordinates to retrieve weather information for the cities of interest

# Define key needed for use API (obtained after registration)
API_key = open("access_keys/API_Key_OpenWeather.txt", "r").read()

#Latitude and longitude form df as lists
lat=GPS_df["lat"].values.tolist()
lon=GPS_df["lon"].values.tolist()

#List comprehension to get request for all of the cities found (by latitude and longitude)
api_weath=[req_weather_fn(x,y, key=API_key) for x,y in zip(lat,lon)]

print('API weather results length: ', len(api_weath)) 
print('API weather keys length: ',len(api_weath[0])) 
api_weath[0]

# Create dataframe from request response
weather_df_raw = req_to_df_fn(api_weath)
weather_df_raw.head()

#Drop current weather information. Keep only daily
weather_df=weather_df_raw.copy().reset_index(drop=True)
weather_df.drop('current',axis=1,inplace=True)
print(weather_df.shape)

# First data explosion: on chosen column: daily
df_explode = explode_df_fn(weather_df, 'daily')
print(df_explode.shape)
df_explode.head()

# Second data explosion: explode weather info for every day
names_7days=[str(df_explode.columns[x]) for x in range(5,12)]
td=["Today"]
days_ls=td+names_7days

weather_perday_df = explodeall_df_fn(days_ls, df_exploded=df_explode)
weather_perday_df.head(44)

# Export weather df
weather_perday_df.to_csv('data/processed/top_35_cities_france_weather_7days.csv', index=False)

## 3. Final dataset: Expanded weather data for 7 days for 35 cities

City Name included with lat and lon  
1 row equivalent to 1 day weather info for a specific city

# Merge GPS_df to weather_perday_df to get name of city

# Before merging: update type of columns in GPS df
GPS_df['lat']=pd.to_numeric(GPS_df["lat"])
GPS_df['lon']=pd.to_numeric(GPS_df["lon"])

# Round lat and lon to signif figures for merging
GPS_rounded_df=GPS_df.copy()
GPS_rounded_df['lat']=[np.round(num,4) for num in GPS_rounded_df['lat']]
GPS_rounded_df['lon']=[np.round(num,4) for num in GPS_rounded_df['lon']]

GPS_rounded_df.to_csv('data/processed/top_35_cities_france_gpscoord.csv', index=False)

GPS_rounded_df.head()

# Final, merged dataframe. Lost info for Collioure despite updating on latitude and longitude 
gps_7dweather_df=weather_perday_df.merge(GPS_rounded_df,left_on=["lat","lon"],right_on=["lat","lon"])
print(gps_7dweather_df.shape)
print('{} unique cities: {}'.format(gps_7dweather_df.Search_Name.nunique(), gps_7dweather_df.Search_Name.unique()))
gps_7dweather_df.head()

# Export final dataframe with weather for 7 days for top french cities to visit
gps_7dweather_df.to_csv('results/top_35_cities_france_weather_gps.csv', index=False)
