==> 01-Kayak/02_booking_scrap1.ipynb <==
## Hotel info: scraping booking.com for top 35 cities to visit in France  

In this notebook I :  
-  Retrieved information for the top 100 hotels found in the [top 35 best cities in France](https://one-week-in.com/35-cities-to-visit-in-france/)
-  Hotel information is scraped from [booking.com](https://www.booking.com)

## Import libraries

import pandas as pd
import numpy as np
import os
import logging

import scrapy
from scrapy.crawler import CrawlerProcess

from src.booking_scrap1 import * # import spiders and list of top 35 cities

## Top 35 cities  
Check list of names of cities without french accents


# Cities_meta_df imported in src.booking_scrap1 from src.cities_weather
top_cities_weather_ls = cities_meta_df['cities_clean'].tolist()

print(len(top_cities_weather_ls))
top_cities_weather_ls[:5]

## Scrap  hotel information  

Carry out two rounds of scraping to obtain the following information for each hotel found in each city:  

*Scrap1*
*   **hotel_name** : hotel name  
*   **suburbs** : if found, indicate city suburb in which hotel is found  
*   **link** : url to its booking.com page  
*   **rating** : score given by the website users  
*   **room_type** : description of available room  
*   **price** : price found for the indicated room and length of stay  
*   **stay** : length of stay chosen in search  
*   **guests** : number of guests (adults) chosen in search  
*   **description** : text description of the hotel     
*   **location** : distance of location to city centre  
*   **map_link** : url to its location on a map  


*Scrap2*  
* hotel name  
* hotel latitude and longitude

### Scrap1  
Following an update of booking.com, could no longer find pagination link  
Solution: create 4 spiders to scrap top 100 best hoteks (booking default ranking)

# Name of the file where the results will be saved
filename = "scrap1_hotels_topcities_booking.csv"

# If file already exists, delete it before crawling (because Scrapy will concatenate the last and new results otherwise)
if filename in os.listdir('data/interim/'):
        os.remove('data/interim/' + filename)


process = CrawlerProcess(settings = {
    'USER_AGENT': 'Chrome/97.0 (compatible; MSIE 7.0; Windows NT 5.1)',
    'LOG_LEVEL': logging.INFO,
    "FEEDS": {
        'data/interim/' + filename: {"format": "csv"},
    }
})

# Start the crawling using the spider you defined above
process.crawl(HotelBookingSpider)
process.crawl(HotelBookingSpiderp2)
process.crawl(HotelBookingSpiderp3)
process.crawl(HotelBookingSpiderp4)
process.start()

### Scraping output

booking_df = pd.read_csv('data/interim/scrap1_hotels_topcities_booking.csv')
print(booking_df.shape)
booking_df.head()

# Check for unique city values: 
print(booking_df.city.unique().shape)
print(booking_df.city.unique())


# Clean for cases where there is the concatenation line 'city' and the 0 properties found
booking_df_clean = booking_df.copy()

booking_df_clean = booking_df_clean.drop(booking_df_clean[booking_df_clean.city == 'city'].index)
booking_df_clean = booking_df_clean.drop(booking_df_clean[booking_df_clean.city == '0 properties are available in and around this destination'].index)
booking_df_clean.shape

# Check for unique city values: 
print(booking_df_clean.city.unique().shape)
print(booking_df_clean.city.unique())

booking_df_clean.to_csv('data/processed/scrap1_hotels_topcities_booking-clean.csv', index = False)

## Scrap2  

Go to notebook 02_booking_scrap2.ipynb for scraping of hotel latitude and longitude
