==> 01-Kayak/02_booking_scrap2.ipynb <==
## Hotel location: scraping booking.com for location of hotels in top 35 cities to visit in France  

In this notebook I :  
-  Scraped the latitude and longitude for all the top 100 hotels found in the [top 35 best cities in France](https://one-week-in.com/35-cities-to-visit-in-france/)
-  Hotel latitude and longitude is scraped from [booking.com](https://www.booking.com), using the link to the map of each hotel

- For scraping of hotel information, see notebook 02-booking_scrap1.ipynb

## Import libraries

import pandas as pd
import os
import logging

from bs4 import BeautifulSoup

import scrapy
from scrapy.crawler import CrawlerProcess

from src.booking_scrap2 import *

## Scrap2  

Use map links for each hotel retrieved from booking.com to obtatin latitude and longitude of hotel

from src.booking_scrap1 import hotels_links_ls

# Name of the file where the results will be saved
filename = "scrap2_hotels_topcities_booking.csv"


# If file already exists, delete it before crawling (because Scrapy will concatenate the last and new results otherwise)
if filename in os.listdir('data/processed/'):
        os.remove('data/processed/' + filename)


process = CrawlerProcess(settings = {
    'USER_AGENT': 'Chrome/97.0 (compatible; MSIE 7.0; Windows NT 5.1)',
    'LOG_LEVEL': logging.INFO,
    "FEEDS": {
        'data/processed/' + filename: {"format": "csv"},
    }
})

# Start the crawling using the spider you defined above
process.crawl(HotelLatLonSpider)
process.start()

### Scraping output

booking_df_loc = pd.read_csv('data/processed/scrap2_hotels_topcities_booking.csv')
print(booking_df_loc.shape)
booking_df_loc.head()

## Merge Scrap 1 and Scrap2

# Import scrap1 result
booking_df = pd.read_csv('data/processed/scrap1_hotels_topcities_booking-clean.csv')
booking_df.columns

# Check hotel names match
ls_hot =booking_df['hotel_name'].unique().sort()
ls_loc = booking_df_loc['hotel_name'].unique().sort()
ls_hot == ls_loc

# Merge dfs
booking_df_s3 = booking_df.merge(booking_df_loc, on=['hotel_name'])
print(booking_df_s3.shape)
print(booking_df_s3.columns)
booking_df_s3.head()

# Output merged dataframe to be uploaded to S3 bucket
booking_df_s3.to_csv('results/scrap_booking_s3.csv', index= False)
