==> 01-Kayak/03_s3_db.ipynb <==
## S3 bucket: store hotel and weather info for top 35 cities to visit in France  

In this notebook I :  
-  Created an **AWS s3 bucket** containing a single dataframe with weather and hotel information for the [top 35 best cities in France](https://one-week-in.com/35-cities-to-visit-in-france/)
-  City gps and weather and information was scrapped from the [nominatim](https://nominatim.org/) and [openweathermap](https://openweathermap.org/api/one-call-api) APIs
-  Hotel information was scraped from [booking.com](https://www.booking.com)

## Import libraries

import boto3
import pandas as pd
import json
import ast

## AWS configuration

from src.credentials import keys_aws

#Create session using AWS access keys and IDs
session = boto3.Session(aws_access_key_id=keys_aws.loc[0,'Access key ID'], 
aws_secret_access_key=keys_aws.loc[0,'Secret access key'])

## Ressource session

s3_resource = session.resource('s3')
bucket_name = "kayak-amorenov"

## Create bucket

# Indicate public-read for read access to AllUsers group
kayak_bucket = s3_resource.create_bucket(Bucket=bucket_name,
ACL = 'public-read',
CreateBucketConfiguration={
        'LocationConstraint': 'eu-west-3'})

kayak_bucket

# Upload bucket policy to make S3 bucket publicly available
file = open("config/AWS_bucket_policy.txt", "r")

policy_file = file.read()
policy_dict = ast.literal_eval(policy_file)

policy_dict

# Indicate bucket name inside policy
policy_dict['Statement'][0]['Resource'][0] = 'arn:aws:s3:::{}/*'.format(bucket_name)

policy_dict

# Update bucket policy 
bucket_policy = s3_resource.BucketPolicy(bucket_name)

bucket_policy_new = json.dumps(policy_dict)
bucket_policy.put(Policy=bucket_policy_new)

## Import data for bucket

# Import output from weather scraping
weather_df_s3 = pd.read_csv('results/top_35_cities_france_weather_gps.csv')
print(weather_df_s3.shape)
weather_df_s3.head()

# Import output from hotels  scraping
booking_df_s3 = pd.read_csv('results/scrap_booking_s3.csv')
print(booking_df_s3.shape)
booking_df_s3.head()

# Homogenize city names to be able to merge dataframes if necessary
from src.cities_weather import cities_meta_df
print(cities_meta_df.shape)
print(cities_meta_df.head())

weather_df_s3.rename(columns={"Search_Name" : "city"}, inplace=True)
booking_df_s3.rename(columns={"city":"city_clean"}, inplace=True)

weather_df_s3_clean = weather_df_s3.merge(cities_meta_df, how='left', on='city')
booking_df_s3_clean = booking_df_s3.merge(cities_meta_df, how='left', on='city_clean')

print(weather_df_s3_clean.columns)
print(booking_df_s3_clean.columns)

## Upload data to bucket

# Upload file to bucket
weathercsv_s3 = weather_df_s3_clean.to_csv(index=False)
booking_csv_s3 = booking_df_s3_clean.to_csv(index=False)

put_object = kayak_bucket.put_object(Key="top35cities_france_scrap_weather7d.csv", Body=weathercsv_s3)
put_object = kayak_bucket.put_object(Key="top35cities_france_scrap_hotelsbooking.csv", Body=booking_csv_s3)

## Verify public access to bucket

# Verify public access to created bucket: download file
# Additionally checked on own aws account (see screenshot of aws in reports/figures/01-Kayak_aws_bucket_creation.png)

test_bucket_weather = pd.read_csv('https://kayak-amorenov.s3.eu-west-3.amazonaws.com/top35cities_france_scrap_weather7d.csv')
test_bucket_booking = pd.read_csv('https://kayak-amorenov.s3.eu-west-3.amazonaws.com/top35cities_france_scrap_hotelsbooking.csv')

print(weather_df_s3_clean.shape)
print(test_bucket_weather.shape)


print(booking_df_s3_clean.shape)
print(test_bucket_booking.shape)


test_bucket_weather.head()

test_bucket_booking.head()
